Debug to get orchestrator functional for sentiment task
